deployment/requirements.txt
scikit_learn
pandas
matplotlib
seaborn
Pillow